Begin working on the AppBarLayout to emulate what Google Calendar and Outlook Web App use.
Using a CollapsingToolbarLayout that will contain the calendar view within it. This will collapse and expand as needed. For now, a pretty image of a butterfly is used to get the general functionality of the collapsing and expanding working.

The CollapsingToolbarLayout has a

Within the CoordinatorLayout, a NestedScrollView class is added with 20 buttons to create a large list. When the user scrolls this list, the toolbar layout will expand and collapsed based on the XML attributes given to it.

The layout_scrollFlags attribute is set to scroll|exitUntilCollapsed that will collapse the toolbar when the user scrolls down in the NestedScrollView. However, by default this will also open the toolbar when the user scrolls to the top of the scroll view. This is not desired and a custom attribute named behavior_scrollTopAction is defined to disable this behavior. The functionality for this fix is coded in the CalendarScrollingBehavior attached to the AppBarLayout via the layout_behavior attribute.

Additionally, the user can drag the toolbar down to get the collapsing toolbar layout to expand. But, this only works if the user if at the top of the nested scroll view. Since we will typically want the user to expand the toolbar layout to show the calendar at any time, a custom attribute named behavior_dragType is added to the AppBarLayout. This functionality is coded via the CalendarScrollingBehavior class.