Finish working on switching between toolbars at run-time
There is a menu option that allows you to switch between two toolbars currently. One has a title and subtitle where the subtitle is the month and year while the second one is just a title that says the month and year.

Clean up toolbars by renaming items by a set standard and removing any unnecessary code.

Create AppBarLayout in library that extends AppBarLayout. All it does is contain a state variable that keeps track of whether the layout is expanded or collapsed. This is done through the onOffsetChangedListener.

Add functionality to the date selector button. Clicking it will expand or collapse the toolbar