Added support for selecting a day from the calendar
Selecting any of the days in the calendar will add a circular span on that day that is the color of the accent color. Currently, the color or size cannot be configured but that will change in the future.

In addition, the current day also has a circular span of the same color to show what today is.

Added a listener that allows the developer to define their custom logic for each day of the calendar. They display whatever spans they like in this listener.