Create dimens.xml file in sample activity to store the padding for the circles that appear over CalendarView days when the item is selected.
Add listener in sample activity when a new MonthView is created. A listener for each MonthView is attached in this function to callback a spannable function when the calendar day text view is being drawn. The developer can set any special effects on the spannable text in this callback and then it will be shown in that regard.

Make the toolbar pretty. This involved making the calendar background white which is a similar color as the main screen background. Thus, to establish it on a new level, elevation was added to it. This is only supported from API 21+ and I need to support for the shadows (through a different method) for API 15-20.

Setup the selected circle and today circle on the month view. They are variants of the primary color but one is tinted and the other is not.