Remove EditActivity since it was not doing anything. Just a remnant from when I imported some of the project files.
Remove image_pretty since I have finished working on the collapsing toolbar and now I have an actual calendar view to put there.

Create on selected day listeners to the calendar view and month view class. The calendar view class simply hooks the current month view listener.
Note: I just made the discovery and realization that there MUST ALWAYS be a selected day. Because with the agenda, it cannot be nothing. Just on startup, the selected day is today which has a different icon.

Continue working on Coordinator class that synchronizes the data between the toolbar title view and calendar view. (In the future, this will synchronize the agenda view and such too)

Coordinator creates a listener to CalendarView on selected day change that will update the title.